+++ CHANGELOG.md
@@ -6,8 +6,10 @@ This project adheres to `Semantic Versioning <https://semver.org/>`_.
 ## [Unreleased]
 
 ### Added
+- FlatBuffers schemas and package dependency for UCX bridge message types
 
 ### Changed
+- UCX publish and subscribe nodes now use the Isaac OS FlatBuffers wire format
 
 ### Removed